+++ src/components/CookieNotice/CookieNotice.test.tsx
