+++ src/components/Hints/Hints.test.tsx
