+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import { NotificationProvider } from './Notification';
+import { NotificationProvider } from './components/Notification/Notification';
 
 const root = ReactDOM.createRoot(
     document.getElementById('root') as HTMLElement
